pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "pptx-translate"
version = "0.1.0"
description = "Translate PowerPoint .pptx files while preserving layout and formatting."
readme = "README.md"
requires-python = ">=3.10"
license = { text = "MIT" }
authors = [{ name = "Your Name" }]
dependencies = [
    "openai>=1.30.0",
    "python-pptx>=0.6.23",
    "Pillow>=10.0.0",
]

[project.scripts]
pptx-translate = "pptx_translate.cli:main"

[tool.setuptools.packages.find]
where = ["src"]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
    "python-pptx>=0.6.23",
]
ocr = [
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
]
